+++ src/comments/services/comments.service.ts
@@ -1,11 +1,11 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { Comment } from 'src/comments/types/comments';
 import { PrismaService } from 'src/database/prisma.service';
-import { UserService } from 'src/user/services/user.service';
-import { ProjectsService } from 'src/projects/services/projects/projects.service';
 import { Comments } from '@prisma/client';
 import { LoggerService } from 'src/logger/logger.service';
 import { i18n } from 'src/i18n';
+import { ProjectsService } from 'src/projects/services/projects.service';
+import { UserService } from 'src/user/services/user.service';
 
 @Injectable()
 export class CommentsService {
