+++ src/decisions/services/decisions.service.ts
@@ -1,11 +1,11 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/database/prisma.service';
 import { Decision } from 'src/decisions/types/decision';
-import { ProjectsService } from 'src/projects/services/projects/projects.service';
 import { UserService } from 'src/user/services/user.service';
 import { Decisions } from '@prisma/client';
 import { LoggerService } from 'src/logger/logger.service';
 import { i18n } from 'src/i18n';
+import { ProjectsService } from 'src/projects/services/projects.service';
 
 @Injectable()
 export class DecisionsService {
