+++ src/follow-project/service/follow-project.service.ts
@@ -3,7 +3,7 @@ import { PrismaService } from 'src/database/prisma.service';
 import { Data } from 'src/follow-project/types/follow';
 import { i18n } from 'src/i18n';
 import { LoggerService } from 'src/logger/logger.service';
-import { ProjectsService } from 'src/projects/services/projects/projects.service';
+import { ProjectsService } from 'src/projects/services/projects.service';
 import { UserService } from 'src/user/services/user.service';
 
 @Injectable()
